Refactor javascript for roll handling in new view
Roll handling scripts were spread across a number of obtuse functions. Refactored into more logical functions and fixed a bug with selecting a roll in a selector that already had a value rather than clearing the value first. This cause the original value in the selector to be removed from the list until a value was cleared in any field. Also changed the rolls list to surround a selected value with the <del> tag rather than removing it from the list. This will allow more versatility when displaying or hiding selected rolls going forward.